Add additional values into proto file
Related https://github.com/the-cloudkeeper-project/cloudkeeper-one/issues/40

This adds 2 values that were requested in mentioned issue. In addition,
we decided to rework attributes, since they are not used. Instead of
attributes we use digest(hash) for comparison of individual images.
This digest should be SHA512, but can be any other hash algorithm.
Digests are calculated for appliance and for image separately, since
both are used for different purpose.